+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# [1.2.3-3-final]
+
+- Fixed departure airport displaying in Proc tab when arrival should be displ,ayed
+- Added beluxvacc.org, portugal-vacc.org, vathk.com, hq.vat-sea.com to whitelist
+
 # [1.2.3-2-final]
 
 - Fixed **some** cyrillic ATC info not being parsed correctly